+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # IDLeDOM x.x.x (not yet released)
+* Add DOMException interface.
 
 # IDLeDOM 0.4.0 (2021-05-11)
 * Add interfaces for HTML IDL, in particular the HTML*Element classes.